add page_ctl into tiflash binary
