--- dbms/src/Storages/DeltaMerge/tools/page_ctl/Main.cpp
@@ -1,3 +1,0 @@
-//
-// Created by tongli chen on 2022/5/24.
-//
