+++ dbms/src/Storages/Page/tools/page_ctl/Main.cpp
@@ -1,0 +1,9 @@
+//
+// Created by tongli chen on 2022/5/24.
+//
+#include <Storages/Page/tools/page_ctl/PageStorageCtl.h>
+
+int main(int argc, char ** argv)
+{
+    return DB::PageStorageCtl::mainEntry(argc, argv);
+}
