+++ dbms/src/Storages/Page/tools/page_ctl/PageStorageCtl.h
@@ -1,0 +1,14 @@
+//
+// Created by tongli chen on 2022/5/25.
+//
+
+#pragma once
+
+namespace DB
+{
+class PageStorageCtl
+{
+public:
+    static int mainEntry(int argc, char ** argv);
+};
+} // namespace DB
