+++ dbms/src/Storages/Page/tools/page_ctl/page_storage_ctl_v2.h
@@ -1,0 +1,6 @@
+//
+// Created by tongli chen on 2022/5/25.
+//
+#pragma once
+
+int pageStorageV2CtlEntry(int argc, char ** argv);
