+++ dbms/src/Storages/Page/tools/page_ctl/page_storage_ctl_v3.cpp
@@ -12,21 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <Common/UnifiedLogPatternFormatter.h>
 #include <Encryption/MockKeyManager.h>
 #include <Interpreters/Context.h>
 #include <Poco/ConsoleChannel.h>
-#include <Poco/FormattingChannel.h>
-#include <Poco/Logger.h>
 #include <Poco/PatternFormatter.h>
-#include <Poco/Runnable.h>
-#include <Poco/ThreadPool.h>
-#include <Poco/Timer.h>
-#include <Server/DTTool/DTTool.h>
+#include <Server/CLIService.h>
 #include <Storages/Page/V3/PageDirectory.h>
 #include <Storages/Page/V3/PageDirectoryFactory.h>
 #include <Storages/Page/V3/PageStorageImpl.h>
-#include <Storages/Page/WriteBatch.h>
 #include <Storages/PathPool.h>
 #include <TestUtils/MockDiskDelegator.h>
 
@@ -34,7 +27,9 @@
 
 namespace DB::PS::V3
 {
-
+extern "C" {
+void run_raftstore_proxy_ffi(int argc, const char * const * argv, const DB::EngineStoreServerHelper *);
+}
 struct ControlOptions
 {
     enum DisplayType
@@ -71,9 +66,7 @@ ControlOptions ControlOptions::parse(int argc, char ** argv)
         ("enable_fo_check,E", value<bool>()->default_value(true), "Also check the evert field offsets. This options only works when `display_mode` is 4.") //
         ("query_ns_id,N", value<UInt64>()->default_value(DB::TEST_NAMESPACE_ID), "When used `check_page_id`/`query_page_id`/`query_blob_id` to query results. You can specify a namespace id.")("check_page_id,C", value<UInt64>()->default_value(UINT64_MAX), "Check a single Page id, display the exception if meet. And also will check the field offsets.") //
         ("query_page_id,W", value<UInt64>()->default_value(UINT64_MAX), "Quert a single Page id, and print its version chaim.") //
-        ("query_blob_id,B", value<UInt32>()->default_value(UINT32_MAX), "Quert a single Blob id, and print its data distribution.")
-        ("imitative,I", value<bool>()->default_value(true), "Use imitative context instead. (encryption is not supported in this mode so that no need to set config_file_path)")
-        ("config_file_path,C", value<std::string>(), "Path to TiFlash config (tiflash.toml).");
+        ("query_blob_id,B", value<UInt32>()->default_value(UINT32_MAX), "Quert a single Blob id, and print its data distribution.")("imitative,I", value<bool>()->default_value(true), "Use imitative context instead. (encryption is not supported in this mode so that no need to set config_file_path)")("config_file_path", value<std::string>(), "Path to TiFlash config (tiflash.toml).");
 
 
     static_assert(sizeof(DB::PageId) == sizeof(UInt64));
@@ -105,17 +98,20 @@ ControlOptions ControlOptions::parse(int argc, char ** argv)
     opt.check_page_id = options["check_page_id"].as<UInt64>();
     opt.query_ns_id = options["query_ns_id"].as<UInt64>();
     opt.is_imitative = options["imitative"].as<bool>();
-    if (opt.is_imitative && options.count("config-file") != 0)
+    if (opt.is_imitative && options.count("config_file_path") != 0)
     {
-        std::cerr << "config-file is not allowed in imitative mode" << std::endl;
+        std::cerr << "config_file_path is not allowed in imitative mode" << std::endl;
         exit(0);
     }
-    else if (!opt.is_imitative && options.count("config-file") == 0)
+    else if (!opt.is_imitative && options.count("config_file_path") == 0)
     {
-        std::cerr << "config-file is required in proxy mode" << std::endl;
+        std::cerr << "config_file_path is required in proxy mode" << std::endl;
         exit(0);
     }
-    opt.config_file_path = options["config_file_path"].as<std::string>();
+    if (options.count("config_file_path") != 0)
+    {
+        opt.config_file_path = options["config_file_path"].as<std::string>();
+    }
 
     if (opt.display_mode < DisplayType::DISPLAY_SUMMARY_INFO || opt.display_mode > DisplayType::CHECK_ALL_DATA_CRC)
     {
@@ -127,10 +123,10 @@ ControlOptions ControlOptions::parse(int argc, char ** argv)
     return opt;
 }
 
-class PageStorageControl
+class PageStorageControlV3
 {
 public:
-    explicit PageStorageControl(const ControlOptions & options_)
+    explicit PageStorageControlV3(const ControlOptions & options_)
         : options(options_)
     {
     }
@@ -145,7 +141,7 @@ public:
         else
         {
             PageDirectory::MVCCMapType type;
-            DTTool::CLIService service(getPageStorageV3Info, options, options.config_file_path, DTTool::run_raftstore_proxy_ffi);
+            CLIService service(getPageStorageV3Info, options, options.config_file_path, run_raftstore_proxy_ffi);
             service.run({""});
         }
     }
@@ -176,7 +172,7 @@ private:
         BlobStore::Config blob_config;
 
         PageStorage::Config config;
-        PageStorageImpl ps_v3("PageStorageControl", delegator, config, file_provider_ptr);
+        PageStorageImpl ps_v3("PageStorageControlV3", delegator, config, file_provider_ptr);
         ps_v3.restore();
         PageDirectory::MVCCMapType & mvcc_table_directory = ps_v3.page_directory->mvcc_table_directory;
 
@@ -515,11 +511,5 @@ using namespace DB::PS::V3;
 void pageStorageV3CtlEntry(int argc, char ** argv)
 {
     const auto & options = ControlOptions::parse(argc, argv);
-    PageStorageControl(options).run();
-}
-
-int main(int argc, char ** argv)
-{
-    pageStorageV3CtlEntry(argc, argv);
-    return 0;
+    PageStorageControlV3(options).run();
 }
