+++ dbms/src/Storages/Page/tools/page_ctl/page_storage_ctl_v3.h
@@ -1,0 +1,7 @@
+//
+// Created by tongli chen on 2022/5/25.
+//
+
+#pragma once
+
+void pageStorageV3CtlEntry(int argc, char ** argv);
